Rename: 응집성을 위해서 테스트 서포트에 사용되는 클래스들을 한 패키지 아래로 이동시킴
