+++ src/test/java/com/growup/pms/test/support/WithMockSecurityUserSecurityContextFactory.java
@@ -1,6 +1,7 @@
-package com.growup.pms.test.annotation;
+package com.growup.pms.test.support;
 
 import com.growup.pms.auth.domain.SecurityUser;
+import com.growup.pms.test.annotation.WithMockSecurityUser;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
