+++ src/test/java/com/growup/pms/user/controller/UserControllerV1Test.java
@@ -6,7 +6,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.growup.pms.test.DefaultControllerSliceTest;
+import com.growup.pms.test.support.DefaultControllerSliceTest;
 import com.growup.pms.test.annotation.AutoKoreanDisplayName;
 import com.growup.pms.test.fixture.user.UserCreateRequestFixture;
 import com.growup.pms.user.dto.UserCreateRequest;
